+++ src/main/java/br/com/backEndVendas/service/dao/PedidoDao.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PedidoDao extends JpaRepository<Pedido, Long> {
+public interface PedidoDao extends JpaRepository<Pedido, Integer> {
+
 }
